requirements.txt
python-dotenv
ipython
qdrant_client 
llama-index-vector-stores-qdrant 
llama-index-readers-file 
llama-index-embeddings-ollama
llama-index-embeddings-huggingface
llama_index-vector_stores-chroma 
llama-index-llms-openai 
llama-index 
openai
chromadb
langchain
scrapy
streamlit